REname shape-from-mesh into collision-from-mesh
